move `Just*` from `optype.typing` to `optype` and deprecate
