+++ optype/__init__.pyi
@@ -1,15 +1,4 @@
-from . import (
-    copy,
-    dataclasses,
-    inspect,
-    io,
-    json,
-    numpy,
-    pickle,
-    string,
-    types,
-    typing,
-)
+from . import copy, dataclasses, inspect, io, json, numpy, pickle, string, types, typing
 from ._core._can import (
     CanAEnter,
     CanAEnterSelf,
@@ -322,6 +311,7 @@ from ._core._has import (
     HasTypeParams,
     HasWrapped,
 )
+from ._core._just import Just, JustComplex, JustFloat, JustInt
 
 __all__ = [
     "CanAEnter",
@@ -546,6 +536,10 @@ __all__ = [
     "HasSlots",
     "HasTypeParams",
     "HasWrapped",
+    "Just",
+    "JustComplex",
+    "JustFloat",
+    "JustInt",
     "__version__",
     "copy",
     "dataclasses",
